hardware/requirements.txt
Pillow==5.2.0
PyYAML==5.4
six==1.11.0
websocket-client==0.48.0
